requirements.txt
Django==1.7.1
django-debug-toolbar==1.2.2
django-localflavor==1.1
docutils==0.12
geopy==1.6.1
psycopg2==2.5.4
sqlparse==0.1.14
